feat: epoch component error handling
